Add custom queries for data incident header and update incident table component test
